Renamed lib folder to Domain
