+++ app/javascript/Domain/Entities/AbstractImperialGame.js
