+++ app/javascript/Domain/Entities/Board/Province.js
