+++ app/javascript/Domain/Entities/Entity.js
