+++ app/javascript/Domain/Entities/ImperialAsiaGame.js
