+++ app/javascript/Domain/Entities/ImperialEuropeGame.js
