+++ app/javascript/Domain/Entities/Nations/Nation.js
