+++ app/javascript/Domain/Entities/Rondel.js
