+++ app/javascript/Domain/Entities/RondelSlots/FactorySlot.js
