+++ app/javascript/Domain/Entities/RondelSlots/RondelSlot.js
