+++ app/javascript/Domain/ImperialGameCoordinator.js
