+++ app/javascript/Domain/UseCases/Rondels/MoveToSlot.js
