+++ app/javascript/Domain/UseCases/Rondels/SlotSelection/AvailableSlots.js
