+++ app/javascript/Domain/UseCases/Rondels/SlotSelection/SlotDistanceCosts.js
