+++ app/javascript/Domain/auction.js
