+++ app/javascript/Domain/auction.test.js
