+++ app/javascript/Domain/auction2030Setup.js
