+++ app/javascript/Domain/auctionAsiaSetup.js
