+++ app/javascript/Domain/auctionSetup.js
