+++ app/javascript/Domain/auctionSetup.test.js
