+++ app/javascript/Domain/availableBondPurchases.js
