+++ app/javascript/Domain/board.js
