+++ app/javascript/Domain/board2030.js
