+++ app/javascript/Domain/boardAsia.js
