+++ app/javascript/Domain/constants.js
