+++ app/javascript/Domain/enum.js
