+++ app/javascript/Domain/enum.test.js
