+++ app/javascript/Domain/gameBoard.js
