+++ app/javascript/Domain/schnelleinsteigLog.js
