+++ app/javascript/Domain/schnelleinsteig_legacy_tests.js
