+++ app/javascript/Domain/setOldAuctionState.js
