+++ app/javascript/Domain/standardAsiaSetup.js
