+++ app/javascript/Domain/standardAsiaSetup.test.js
