+++ app/javascript/Domain/standardSetup.js
